--- a/src/Logger/Adapter/Noop.php
+++ b/src/Logger/Adapter/Noop.php
@@ -1,0 +1,42 @@
+<?php
+
+/**
+ * This file is part of the Phalcon Framework.
+ *
+ * (c) Phalcon Team <team@phalcon.io>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Logger\Adapter;
+
+use Phalcon\Logger\Item;
+
+/**
+ * Class Noop
+ *
+ * @package Phalcon\Logger\Adapter
+ */
+class Noop extends AbstractAdapter
+{
+    /**
+     * Closes the stream
+     */
+    public function close(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Processes the message i.e. writes it to the file
+     *
+     * @param Item $item
+     */
+    public function process(Item $item): void
+    {
+        // noop
+    }
+}