… .github/ISSUE_TEMPLATE/bug_report.md
… .github/ISSUE_TEMPLATE/feature_request.md
… .github/issue_template/bug_report.md
… .github/issue_template/feature_request.md
… .github/pull_request_template.md
… .github/stale.yml
… README.md
… codacy.yml
… composer.json
… composer.lock
… src/Logger/AdapterFactory.php
… src/Logger/Exception.php
… src/Logger/Formatter/AbstractFormatter.php
… src/Logger/Formatter/FormatterInterface.php
… src/Logger/Formatter/Json.php
… src/Logger/Formatter/Line.php
… src/Logger/Item.php
… src/Logger/Logger.php
… src/Logger/LoggerFactory.php
… tests/_bootstrap.php
… tests/_ci/functions.php
… tests/_data/fixtures/Traits/LoaderTrait.php
… tests/_data/fixtures/Traits/LoggerTrait.php
… tests/unit/Autoload/Loader/AutoloadCest.php
… tests/unit/Autoload/Loader/GetAddSetClassesCest.php
… tests/unit/Autoload/Loader/GetAddSetExtensionsCest.php
… tests/unit/Autoload/Loader/GetAddSetFilesCest.php
… tests/unit/Autoload/Loader/GetAddSetNamespacesCest.php
… tests/unit/Autoload/Loader/LoadFilesCest.php
… tests/unit/Autoload/Loader/RegisterUnregisterCest.php