@Jeckerson 2020-09-09
Before

```
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
| benchmark   | subject            | set | revs  | its | mem_peak   | best     | mean     | mode     | worst    | stdev   | rstdev | diff    |
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
| LoaderBench | bench__construct   | 0   | 20000 | 10  | 986,288b   | 0.471μs  | 0.516μs  | 0.502μs  | 0.591μs  | 0.036μs | 6.96%  | 1.07x   |
| LoaderBench | benchAddClass      | 0   | 20000 | 10  | 986,328b   | 0.538μs  | 0.575μs  | 0.557μs  | 0.644μs  | 0.033μs | 5.70%  | 1.19x   |
| LoaderBench | benchAddExtension  | 0   | 20000 | 10  | 986,336b   | 0.652μs  | 0.688μs  | 0.666μs  | 0.787μs  | 0.044μs | 6.37%  | 1.42x   |
| LoaderBench | benchAddFile       | 0   | 20000 | 10  | 986,328b   | 0.798μs  | 0.834μs  | 0.810μs  | 0.998μs  | 0.059μs | 7.06%  | 1.73x   |
| LoaderBench | benchAddNamespace  | 0   | 20000 | 10  | 986,336b   | 0.949μs  | 0.989μs  | 0.993μs  | 1.035μs  | 0.026μs | 2.60%  | 2.05x   |
| LoaderBench | benchAutoload      | 0   | 20000 | 10  | 986,328b   | 0.686μs  | 0.709μs  | 0.706μs  | 0.731μs  | 0.014μs | 1.91%  | 1.47x   |
| LoaderBench | benchGetClasses    | 0   | 20000 | 10  | 986,328b   | 0.470μs  | 0.493μs  | 0.495μs  | 0.522μs  | 0.014μs | 2.82%  | 1.02x   |
| LoaderBench | benchGetDebug      | 0   | 20000 | 10  | 986,328b   | 0.462μs  | 0.483μs  | 0.482μs  | 0.506μs  | 0.012μs | 2.57%  | 1.00x   |
| LoaderBench | benchGetExtensions | 0   | 20000 | 10  | 986,336b   | 0.473μs  | 0.491μs  | 0.487μs  | 0.511μs  | 0.012μs | 2.48%  | 1.02x   |
| LoaderBench | benchGetFiles      | 0   | 20000 | 10  | 986,328b   | 0.495μs  | 0.518μs  | 0.511μs  | 0.584μs  | 0.024μs | 4.64%  | 1.07x   |
| LoaderBench | benchGetNamespaces | 0   | 20000 | 10  | 986,336b   | 0.479μs  | 0.541μs  | 0.500μs  | 0.703μs  | 0.070μs | 12.88% | 1.12x   |
| LoaderBench | benchLoadFiles     | 0   | 20000 | 10  | 986,328b   | 0.475μs  | 0.502μs  | 0.489μs  | 0.537μs  | 0.022μs | 4.43%  | 1.04x   |
| LoaderBench | benchSetClasses    | 0   | 20000 | 10  | 986,328b   | 0.624μs  | 0.658μs  | 0.662μs  | 0.690μs  | 0.018μs | 2.67%  | 1.36x   |
| LoaderBench | benchSetExtensions | 0   | 20000 | 10  | 986,336b   | 0.741μs  | 0.807μs  | 0.753μs  | 1.302μs  | 0.165μs | 20.46% | 1.67x   |
| LoaderBench | benchSetFiles      | 0   | 20000 | 10  | 986,328b   | 0.874μs  | 0.910μs  | 0.897μs  | 0.999μs  | 0.036μs | 3.92%  | 1.88x   |
| LoaderBench | benchSetNamespaces | 0   | 20000 | 10  | 986,336b   | 1.067μs  | 1.079μs  | 1.076μs  | 1.107μs  | 0.011μs | 1.01%  | 2.23x   |
| LoaderBench | benchRegister      | 0   | 20000 | 10  | 7,773,280b | 53.204μs | 54.313μs | 53.875μs | 55.946μs | 0.853μs | 1.57%  | 112.42x |
| LoaderBench | benchUnregister    | 0   | 20000 | 10  | 986,328b   | 0.472μs  | 0.497μs  | 0.489μs  | 0.561μs  | 0.025μs | 4.98%  | 1.03x   |
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
```

Now

```
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
| benchmark   | subject            | set | revs  | its | mem_peak   | best     | mean     | mode     | worst    | stdev   | rstdev | diff    |
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
| LoaderBench | bench__construct   | 0   | 20000 | 10  | 986,352b   | 0.469μs  | 0.500μs  | 0.508μs  | 0.520μs  | 0.016μs | 3.18%  | 1.02x   |
| LoaderBench | benchAddClass      | 0   | 20000 | 10  | 986,392b   | 0.538μs  | 0.551μs  | 0.552μs  | 0.570μs  | 0.010μs | 1.78%  | 1.12x   |
| LoaderBench | benchAddExtension  | 0   | 20000 | 10  | 986,400b   | 0.648μs  | 0.667μs  | 0.655μs  | 0.704μs  | 0.020μs | 3.01%  | 1.36x   |
| LoaderBench | benchAddFile       | 0   | 20000 | 10  | 986,392b   | 0.817μs  | 0.833μs  | 0.824μs  | 0.900μs  | 0.024μs | 2.91%  | 1.70x   |
| LoaderBench | benchAddNamespace  | 0   | 20000 | 10  | 986,400b   | 0.967μs  | 0.992μs  | 0.977μs  | 1.038μs  | 0.025μs | 2.49%  | 2.02x   |
| LoaderBench | benchAutoload      | 0   | 20000 | 10  | 986,392b   | 0.687μs  | 0.712μs  | 0.709μs  | 0.739μs  | 0.015μs | 2.10%  | 1.45x   |
| LoaderBench | benchGetClasses    | 0   | 20000 | 10  | 986,392b   | 0.468μs  | 0.494μs  | 0.487μs  | 0.540μs  | 0.020μs | 3.95%  | 1.01x   |
| LoaderBench | benchGetDebug      | 0   | 20000 | 10  | 986,392b   | 0.467μs  | 0.491μs  | 0.478μs  | 0.573μs  | 0.031μs | 6.21%  | 1.00x   |
| LoaderBench | benchGetExtensions | 0   | 20000 | 10  | 986,400b   | 0.478μs  | 0.499μs  | 0.507μs  | 0.525μs  | 0.015μs | 2.93%  | 1.02x   |
| LoaderBench | benchGetFiles      | 0   | 20000 | 10  | 986,392b   | 0.480μs  | 0.514μs  | 0.497μs  | 0.600μs  | 0.037μs | 7.28%  | 1.05x   |
| LoaderBench | benchGetNamespaces | 0   | 20000 | 10  | 986,400b   | 0.483μs  | 0.506μs  | 0.491μs  | 0.557μs  | 0.023μs | 4.63%  | 1.03x   |
| LoaderBench | benchLoadFiles     | 0   | 20000 | 10  | 986,392b   | 0.479μs  | 0.495μs  | 0.494μs  | 0.513μs  | 0.012μs | 2.33%  | 1.01x   |
| LoaderBench | benchSetClasses    | 0   | 20000 | 10  | 986,392b   | 0.620μs  | 0.645μs  | 0.647μs  | 0.675μs  | 0.015μs | 2.40%  | 1.31x   |
| LoaderBench | benchSetExtensions | 0   | 20000 | 10  | 986,400b   | 0.737μs  | 0.763μs  | 0.775μs  | 0.782μs  | 0.017μs | 2.18%  | 1.55x   |
| LoaderBench | benchSetFiles      | 0   | 20000 | 10  | 986,392b   | 0.875μs  | 0.901μs  | 0.888μs  | 0.961μs  | 0.026μs | 2.83%  | 1.83x   |
| LoaderBench | benchSetNamespaces | 0   | 20000 | 10  | 986,400b   | 1.082μs  | 1.101μs  | 1.093μs  | 1.142μs  | 0.018μs | 1.65%  | 2.24x   |
| LoaderBench | benchRegister      | 0   | 20000 | 10  | 7,773,272b | 53.237μs | 54.593μs | 53.690μs | 56.854μs | 1.300μs | 2.38%  | 111.14x |
| LoaderBench | benchUnregister    | 0   | 20000 | 10  | 986,392b   | 0.473μs  | 0.496μs  | 0.482μs  | 0.579μs  | 0.032μs | 6.49%  | 1.01x   |
+-------------+--------------------+-----+-------+-----+------------+----------+----------+----------+----------+---------+--------+---------+
```